requirements.txt
# Common requirements
numpy>=1.19.2
pyserial>=3.5
RPi.GPIO>=0.7.0
smbus2>=0.4.1
logging>=0.5.1.2
opencv-python>=4.5.3.56

# Jetson Nano specific
pyzed>=3.5
torch>=1.9.0  # For future ML implementation

# Raspberry Pi specific
adafruit-circuitpython-pca9685>=3.3.4
adafruit-circuitpython-servokit>=1.3.7

# Development tools
pytest>=6.2.5
black>=21.9b0
pylint>=2.11.1 